Issue #26: Support Eslint Flat Config
