+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "eslint-plugin-no-secrets",
-  "version": "0.9.1",
+  "version": "1.0.0-eslint9",
   "description": "An eslint rule that searches for potential secrets/keys in code",
   "main": "index.js",
   "scripts": {
@@ -22,15 +22,16 @@
   "license": "MIT",
   "devDependencies": {
     "eslint": "^7.19.0",
-    "eslint-plugin-jsonc": "^0.9.0",
+    "eslint-plugin-jsonc": "^2.15.1",
     "eslint-plugin-self": "^1.2.0",
     "eslint6": "npm:eslint@^6.8.0",
-    "mocha": "^7.1.1"
+    "eslint8": "npm:eslint@^8.57.0",
+    "eslint9": "npm:eslint@^9.0.0",
+    "mocha": "^10.4.0"
   },
   "peerDependencies": {
     "eslint": ">=3.0.0"
   },
-  "dependencies": {},
   "engines": {
     "npm": ">=6.9.0",
     "node": ">=10.0.0"
